Replace the lon/lat coordinate with geometry coordinates
The feature previously contained a "lon" and "lat" field, which represented the longitude and latitude coordinates of the feature in WGS84 (EPSG:4326). However, the dictionary also included a "geometry" field with the same information, but in a different coordinate reference system (EPSG:28992). This commit replaces the use of the "lon" and "lat" fields with the use of the "geometry" field, transformed to WGS84 (EPSG:4326).